통계 랭킹 admin -> user(allowany) 기능으로 변경
