+++ config/urls.py
@@ -29,7 +29,7 @@ urlpatterns = [
     path("api/cart/", include("apps.carts.urls")),
     path("api/payments/", include("apps.payments.urls")),
     path("api/admin/payments/", include("apps.payments.admin_urls")),
-    path("api/admin/stats/", include("apps.stats.urls")),
+    path("api/", include("apps.stats.urls")),
     path("api/support/", include("apps.support.urls")),
     path("api/admin/support/", include("apps.support.admin_urls")),
     # Swagger 문서
